requirements.txt
dj-database-url==0.5.0
Django==1.9
django-widget-tweaks==1.4.1
djangorestframework==3.2.0
gunicorn==19.10.0
json2html==1.2.1
psycopg2==2.8.5
python-decouple==3.3
PyYAML==3.11
whitenoise==3.2.2
